DBC22-3396: route switch and loading fixes
