+++ src/frontend/src/Components/map/MapWrapper.js
@@ -1,20 +1,23 @@
 // React
-import React, { useEffect, useRef, useState, useCallback } from 'react';
+import React, {useEffect, useRef, useState, useCallback, useContext} from 'react';
 
 // Redux
 import { memoize } from 'proxy-memoize';
 import { useSelector, useDispatch } from 'react-redux';
 
-// Components and functions
+// Internal imports
+import { MapContext } from "../../App";
 import { NetworkError, ServerError } from '../data/helper';
 import * as dataLoaders from './dataLoaders'
 import * as slices from '../../slices';
-
 import DriveBCMap from './Map';
 import PollingComponent from '../shared/PollingComponent';
 
 export default function MapWrapper(props) {
   /* Setup */
+  // Context
+  const { mapContext } = useContext(MapContext);
+
   // Redux
   const dispatch = useDispatch();
   const {
@@ -22,8 +25,8 @@ export default function MapWrapper(props) {
       cameras: { list: cameras, filterPoints: camFilterPoints },
       events: { list: events, filterPoints: eventFilterPoints },
       ferries: { list: ferries, filterPoints: ferryFilterPoints },
-      weather: { filteredList: filteredCurrentWeathers, filterPoints: currentWeatherFilterPoints },
-      regional: { filteredList: filteredRegionalWeathers, filterPoints: regionalWeatherFilterPoints },
+      weather: { list: currentWeathers, filterPoints: currentWeatherFilterPoints },
+      regional: { list: regionalWeathers, filterPoints: regionalWeatherFilterPoints },
       hef: { list: hef, filteredList: filteredHef },
       restStops: { list: restStops, filterPoints: restStopFilterPoints },
     },
@@ -52,6 +55,19 @@ export default function MapWrapper(props) {
   const [showNetworkError, setShowNetworkError] = useState(false);
   const [showServerError, setShowServerError] = useState(false);
 
+  const getInitialLoadingLayers = () => {
+    return {
+      cameras: mapContext.visible_layers.highwayCams,
+      events: mapContext.visible_layers.closures || mapContext.visible_layers.majorEvents ||
+        mapContext.visible_layers.minorEvents || mapContext.visible_layers.roadConditions ||
+        mapContext.visible_layers.futureEvents || mapContext.visible_layers.chainUps,
+      ferries: mapContext.visible_layers.inlandFerries,
+      weathers: mapContext.visible_layers.weather,
+      restStops: mapContext.visible_layers.restStops
+    };
+  };
+  const [loadingLayers, setLoadingLayers] = useState(getInitialLoadingLayers());
+
   // Refs
   const workerRef = useRef(null);
   const isInitialLoad = useRef(true);
@@ -73,17 +89,13 @@ export default function MapWrapper(props) {
 
     loadData();
 
-    console.log(selectedRoute);
-  }, [selectedRoute]);
-
-  useEffect(() => {
     // Cleanup function to terminate the worker when the component unmounts
     return () => {
       if (workerRef.current) {
         workerRef.current.terminate();
       }
     };
-  }, []);
+  }, [selectedRoute]);
 
   const resetWorker = () => {
     // Terminate the current worker if it exists
@@ -114,20 +126,24 @@ export default function MapWrapper(props) {
 
     const routeData = selectedRouteRef.current && selectedRouteRef.current.routeFound ? selectedRouteRef.current : null;
 
-    if (isInitialLoad.current) {
-      dataLoaders.loadCameras(routeData, cameras, null, camFilterPoints, dispatch, displayError, workerRef.current);
-      dataLoaders.loadEvents(routeData, events, null, eventFilterPoints, dispatch, displayError, workerRef.current, isInitialLoad.current, trackedEventsRef);
-      dataLoaders.loadFerries(routeData, ferries, null, ferryFilterPoints, dispatch, displayError, workerRef.current);
-      dataLoaders.loadCurrentWeather(routeData, null, filteredCurrentWeathers, currentWeatherFilterPoints, dispatch, displayError, workerRef.current);
-      dataLoaders.loadRegionalWeather(routeData, null, filteredRegionalWeathers, regionalWeatherFilterPoints, dispatch, displayError, workerRef.current);
-      dataLoaders.loadHef(routeData, hef, filteredHef, null, dispatch, displayError, workerRef.current);
-      dataLoaders.loadRestStops(routeData, restStops, null, restStopFilterPoints, dispatch, displayError, workerRef.current);
-      dataLoaders.loadAdvisories(routeData, advisories, null, advisoryFilterPoints, dispatch, displayError, workerRef.current);
+    setLoadingLayers({
+      cameras: true,
+      events: true,
+      ferries: true,
+      weathers: true,
+      restStops: true
+    });
 
-      isInitialLoad.current = false;
-    } else {
-      dataLoaders.loadEvents(routeData, events, null, eventFilterPoints, dispatch, displayError, workerRef.current, isInitialLoad.current, trackedEventsRef);
-    }
+    dataLoaders.loadCameras(routeData, cameras, null, camFilterPoints, dispatch, displayError, workerRef.current);
+    dataLoaders.loadEvents(routeData, events, null, eventFilterPoints, dispatch, displayError, workerRef.current, isInitialLoad.current, trackedEventsRef);
+    dataLoaders.loadFerries(routeData, ferries, null, ferryFilterPoints, dispatch, displayError, workerRef.current);
+    dataLoaders.loadCurrentWeather(routeData, currentWeathers, null, currentWeatherFilterPoints, dispatch, displayError, workerRef.current);
+    dataLoaders.loadRegionalWeather(routeData, regionalWeathers, null, regionalWeatherFilterPoints, dispatch, displayError, workerRef.current);
+    dataLoaders.loadHef(routeData, hef, filteredHef, null, dispatch, displayError, workerRef.current);
+    dataLoaders.loadRestStops(routeData, restStops, null, restStopFilterPoints, dispatch, displayError, workerRef.current);
+    dataLoaders.loadAdvisories(routeData, advisories, null, advisoryFilterPoints, dispatch, displayError, workerRef.current);
+
+    isInitialLoad.current = false;
   };
 
   return (
@@ -138,7 +154,10 @@ export default function MapWrapper(props) {
         mapProps={props}
         showNetworkError={showNetworkError}
         showServerError={showServerError}
-        trackedEventsRef={trackedEventsRef} />
+        trackedEventsRef={trackedEventsRef}
+        loadingLayers={loadingLayers}
+        setLoadingLayers={setLoadingLayers}
+        getInitialLoadingLayers={getInitialLoadingLayers} />
     </React.Fragment>
   );
 }
